+++ CHANGELOG.md
@@ -7,6 +7,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2022-10-18
+
+## Added
+
+- configuration default values
+
+### Changed
+
+- `grid` pattern `count` min value is `1`
+
+## [0.3.0] - 2022-10-09
+
+### Added
+
+- `mono` and `backward` palette colors order
+- `triangle` pattern has type: `random` or `equilateral`
+
+## Changed
+
+- `straight` is `forward` palette colors order
+
 ## [0.2.0] - 2022-08-15
 
 ### Added
@@ -23,6 +44,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - A lot of features
 
-[Unreleased]: https://github.com/vikian050194/warp/compare/v0.2.0...HEAD
-[0.2.0]: https://github.com/vikian050194/warp/compare/v0.1.0...v0.2.0
-[0.1.0]: https://github.com/vikian050194/warp/releases/tag/v0.1.0
+[Unreleased]: https://github.com/vikian050194/svg/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/vikian050194/svg/compare/v0.3.0...v0.4.0
+[0.3.0]: https://github.com/vikian050194/svg/compare/v0.2.0...v0.3.0
+[0.2.0]: https://github.com/vikian050194/svg/compare/v0.1.0...v0.2.0
+[0.1.0]: https://github.com/vikian050194/svg/releases/tag/v0.1.0